requirements.txt
apache-beam[gcp]==2.60.0
google-cloud-pubsub==2.29.0

